+++ CHANGELOG.md
@@ -57,6 +57,7 @@
 - [**BREAKING**] **unicorn, cli** configs: [`prefer-top-level-await`](https://github.com/sindresorhus/eslint-plugin-unicorn/blob/HEAD/docs/rules/prefer-top-level-await.md) rule is now off by default and enabled in `cli` config.
 - [**BREAKING**] Start using [`eslint-plugin-jsx-a11y-x`](https://npmjs.com/eslint-plugin-jsx-a11y-x) plugin instead of [`eslint-plugin-jsx-a11y`](https://npmjs.com/eslint-plugin-jsx-a11y) for `jsx-a11y` config as it has much less dependencies.
 - [**BREAKING**] `overrideIgnores` root option removed in favor of `ignores.override`.
+- [**BREAKING**] [`@stylistic/eslint-plugin`](https://npmjs.com/@stylistic/eslint-plugin) rules have moved to a separate `stylistic` config.
 - Added a new root option, `defaultConfigsStatus`, to control what configs are enabled or disabled by default. As a result, shared settings prefix has been changed from `jsx-a11y` to `jsx-a11y-x`.
 - Added the ability to override any of the used plugins via `pluginOverrides` option.
 - Exported `isInCi` helper from `ci-info` package.